+++ bin/shapeshifter.js
@@ -6,109 +6,44 @@
  */
 
 var path = require('path');
-var CommandLine = require('command-line-args');
-var Usage = require('command-line-usage');
+var Vorpal = require('vorpal');
 var Transpiler = require('../lib/Transpiler').default;
-var Shapeshifter = require('../package.json');
 
-// Define options
-var optionList = [
-  {
-    name: 'help',
-    alias: 'h',
-    description: 'Display this help menu.',
-    type: Boolean,
-    defaultValue: false,
-  },
-  {
-    name: 'nullable',
-    alias: 'n',
-    description: 'Mark attributes as nullable by default (recommended). Defaults to false.',
-    type: Boolean,
-    defaultValue: false,
-  },
-  {
-    name: 'compact',
-    alias: 'c',
-    description: 'Compact the output by using Schema#define(). Defaults to false.',
-    type: Boolean,
-    defaultValue: false,
-  },
-  {
-    name: 'indent',
-    description: 'The indentation characters to use. Defaults to 2 space indent.',
-    typeLabel: '[underline]{char}',
-    type: String,
-    defaultValue: '  ',
-  },
-  {
-    name: 'format',
-    description: 'The format to generate. Accepts react, flow, or typescript. Defaults to react.',
-    typeLabel: '[underline]{name}',
-    type: String,
-    defaultValue: 'react',
-  },
-  {
-    name: 'schemas',
-    description: 'Include schema class exports in the output. Defaults to false.',
-    type: Boolean,
-    defaultValue: false,
-  },
-  {
-    name: 'attributes',
-    description: 'Include an attribute list in the schema class export. Defaults to false.',
-    type: Boolean,
-    defaultValue: false,
-  },
-  {
-    name: 'types',
-    description: 'Include type definition exports in the output. Defaults to false.',
-    type: Boolean,
-    defaultValue: false,
-  },
-  {
-    name: 'path',
-    type: String,
-    defaultOption: true,
-  },
-];
+var shapeshifter = new Vorpal();
 
-// Define help menu
-var help = [
-  {
-    header: 'Shapeshifter v' + Shapeshifter.version,
-    content: Shapeshifter.description,
-  },
-  {
-    header: 'Usage',
-    content: '$ shapeshifter [options] [underline]{input} > [underline]{output}',
-  },
-  {
-    header: 'Options',
-    optionList: optionList,
-    hide: 'path'
-  }
-];
+shapeshifter
+  .command('build <source>', 'Transpile source files or folder.')
+  .option('-n, --nullable', 'Mark attributes as nullable by default (recommended). Defaults to false.')
+  .option('-c, --compact', 'Reduce the output of schema ORM definitions. Defaults to false.')
+  .option('--indent <char>', 'The indentation characters to use. Defaults to 2 space indent.')
+  .option('--format <name>', 'The format to generate. Accepts react, flow, or typescript. Defaults to react.', ['react', 'flow', 'typescript'])
+  .option('--schemas', 'Include schema class exports in the output. Defaults to false.')
+  .option('--attributes', 'Include an attribute list in the schema class export. Defaults to false.')
+  .option('--types', 'Include type definition exports in the output. Defaults to false.')
+  .action(function({ options, source }) {
+    return new Transpiler({
+      compact: options.compact || false,
+      defaultNullable: options.nullable || false,
+      includeSchemas: options.schemas || false,
+      includeAttributes: options.attributes || false,
+      includeTypes: options.types || false,
+      indentCharacter: options.indent || '  ',
+      renderer: options.format || 'react',
+    })
+      .transpile(path.normalize(path.join(process.cwd(), source)))
+      .then((output) => {
+        // We need to log the output so that it can be piped
+        this.log(output);
 
-// Parse options
-var options = CommandLine(optionList);
+        return output;
+      })
+      .catch((error) => {
+        // Rudimentary error handling and styling
+        this.log(shapeshifter.chalk.red(error.message));
+      });
+  });
 
-// Show help menu
-if (options.help || !options.path) {
-  Transpiler.output(Usage(help));
-
-// Run transpiler
-} else {
-  new Transpiler({
-    compact: options.compact,
-    defaultNullable: options.nullable,
-    includeSchemas: options.schemas,
-    includeAttributes: options.attributes,
-    includeTypes: options.types,
-    indentCharacter: options.indent,
-    renderer: options.format,
-  })
-    .transpile(path.normalize(path.join(process.cwd(), options.path)))
-    .then(Transpiler.output)
-    .catch(Transpiler.error);
-}
+shapeshifter
+  .delimiter(shapeshifter.chalk.magenta('shapeshifter$ '))
+  .history('shapeshifter')
+  .parse(process.argv);